// src/types.ts
export interface DataType {
    key: React.Key;
    first_name: string;
    last_name: string;
    job_title: string;
    city_name: string;
    state_name: string;
    country_name: string
  }

  export interface IEmpleado {
    id_empelado: number;
    first_name: string;
    last_name: string;
    job_title: string;
    city_name: string;
    state_name: string;
    country_name: string
  }